logo de merliteam lleva al inicio, logos agregados y modificacion de la url para los blogs
